feat: printing lambda functions to STDOUT
Problem: App does not connect to AWS Lambda and print the list of
functions.

App actually doesn't do anything useful, we need a POC that can load AWS
profile from the environment and use that to connect to AWS Lambda and
list the functions.

Don't worry about making it pretty, or even a TUI. Just make it work...

Solution: Used `aws_config` to load AWS profile from environment and
then query and print the list of AWS Lambda functions.

Used `clap` to provide `--profile` and `--region` args to override the
AWS profile in the environment.

Testing: Ran it manually, success!